requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py38_0
asn1crypto=1.3.0=py38_0
blas=1.0=mkl
blinker=1.4=py38_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
cachetools=3.1.1=py_0
certifi=2019.11.28=py38_0
cffi=1.14.0=py38h2e261b9_0
chardet=3.0.4=py38_1003
click=7.0=py_0
cryptography=2.8=py38h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
freetype=2.9.1=h8a8886c_1
google-auth=1.11.2=py_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.27.2=py38hf8bcb03_0
idna=2.8=py38_1000
intel-openmp=2020.0=166
jpeg=9b=h024ee3a_2
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
markdown=3.1.1=py38_0
mkl=2020.0=166
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.0=py38h962f231_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py38hfd86e86_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openssl=1.1.1d=h7b6447c_4
pandas=1.0.1=py38h0573a6f_0
pillow=7.0.0=py38hb39fc2d_0
pip=20.0.2=py38_1
protobuf=3.11.4=py38he6710b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.19=py_0
pyjwt=1.7.1=py38_0
pyopenssl=19.1.0=py38_0
pysocks=1.7.1=py38_0
python=3.8.1=h0371630_1
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=py_0
readline=7.0=h7b6447c_5
requests=2.23.0=py38_0
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
setuptools=46.0.0=py38_0
six=1.14.0=py38_0
sqlite=3.31.1=h7b6447c_0
tensorboard=2.1.0=py3_0
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py38_cu101
urllib3=1.25.8=py38_0
werkzeug=1.0.0=py_0
wheel=0.34.2=py38_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
